refactored code. Created separate components for employee information and employee search form. This simplified the search employee component and made it easier to read. Also changed the conditional statements in search employee component. Each component now has one condition that needs to be met in order for them to be rendered. Also, fixed a bug where the back button was still displaying employee information.
