+++ package.json
@@ -3,8 +3,6 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@emotion/react": "^11.9.3",
-    "@emotion/styled": "^11.9.3",
     "@fortawesome/free-solid-svg-icons": "^6.1.2",
     "@mui/icons-material": "^5.8.4",
     "@mui/material": "^5.8.7",
