+++ src/Dashboard.js
@@ -131,27 +131,11 @@ export default function Dashboard() {
           {['New Fit Test', 'Respirators', 'Companies', 'Employees'].map((text, index) => (
             <ListItem key={text} disablePadding>
               <ListItemButton>
-                {/* <ListItemIcon>
-                  {index % 2 === 0 ? <InboxIcon /> : <MailIcon />}
-                </ListItemIcon> */}
                 <ListItemText primary={text} onClick={()=>handleDrawerClick(text)}/>
               </ListItemButton>
             </ListItem>
           ))}
         </List>
-        {/* <Divider /> */}
-        {/* <List>
-          {['All mail', 'Trash', 'Spam'].map((text, index) => (
-            <ListItem key={text} disablePadding>
-              <ListItemButton>
-                <ListItemIcon>
-                  {index % 2 === 0 ? <InboxIcon /> : <MailIcon />}
-                </ListItemIcon>
-                <ListItemText primary={text} />
-              </ListItemButton>
-            </ListItem>
-          ))}
-        </List> */}
       </Drawer>
       <Main open={open}>
         <DrawerHeader />
