+++ src/components/FitTest/QualitativeFitTestList.js
@@ -200,7 +200,6 @@ export default function QualitativeFitTestList(props) {
   const [orderBy, setOrderBy] = useState('calories');
   const [selected, setSelected] = useState([]);
   const [page, setPage] = useState(0);
-  const [dense, setDense] = useState(false);
   const [rowsPerPage, setRowsPerPage] = useState(10);
   const [showDeleteWarning, setShowDeleteWarning] = useState([false, null]);
 
@@ -247,6 +246,8 @@ export default function QualitativeFitTestList(props) {
   const handleEdit = (row) => {
     props.setSelectedFitTest(row);
     props.setShowQualitativeFitTestEdit(true);
+    props.setShowEmployee(false);
+    props.setShowFitTests(false);
   };
 
   const handleDeleteClick = async (id) => {
@@ -266,10 +267,6 @@ export default function QualitativeFitTestList(props) {
     }
     setRowsPerPage(parseInt(rowSelection, 10));
     setPage(0);
-  };
-
-  const handleChangeDense = (event) => {
-    setDense(event.target.checked);
   };
 
   const isSelected = (name) => selected.indexOf(name) !== -1;
@@ -295,7 +292,7 @@ export default function QualitativeFitTestList(props) {
           <Table
             sx={{ minWidth: 750 }}
             aria-labelledby="tableTitle"
-            size={dense ? 'small' : 'medium'}
+            size={'medium'}
           >
             <EnhancedTableHead
               numSelected={selected.length}
@@ -365,7 +362,7 @@ export default function QualitativeFitTestList(props) {
               {emptyRows > 0 && (
                 <TableRow
                   style={{
-                    height: (dense ? 33 : 53) * emptyRows,
+                    height: (53) * emptyRows,
                   }}
                 >
                   <TableCell colSpan={6} />
@@ -384,10 +381,6 @@ export default function QualitativeFitTestList(props) {
           onRowsPerPageChange={handleChangeRowsPerPage}
         />
       </Paper>
-      <FormControlLabel
-        control={<Switch checked={dense} onChange={handleChangeDense} />}
-        label="Dense padding"
-      />
     </Box>
   );
 }
